[coreVideoCodec] Fix handling of pixel aspect ratio numerator and denominator not fitting into uint8_t
PNG can store dpm (dots per meter) values in its pHYs chunk, usually
exceeding the maximum value representable by uint8_t data type. Libavcodec
interprets dpm values as sample aspect ratio numerator and denominator,
from which we kept the least significant byte only, resulting in a bogus
pixel aspect ratio.

Try to reduce sample aspect ratio provided by libavcodec and modify
getPARWidth() and getPARHeight() functions to return uint32_t.

+++ avidemux_core/ADM_coreVideoCodec/include/ADM_codec.h
@@ -52,8 +52,8 @@ public:
 
   virtual ~decoders() { }
   virtual bool initializedOk(void) { return true; }
-  virtual uint8_t getPARWidth (void) { return 1; }
-  virtual uint8_t getPARHeight (void) { return 1; }
+  virtual uint32_t getPARWidth (void) { return 1; }
+  virtual uint32_t getPARHeight (void) { return 1; }
   virtual bool setParam (void) { return false; }
   virtual bool uncompress (ADMCompressedImage * in, ADMImage * out) = 0;
